yolov8/requirements.txt
ultralytics
torch
torchvision
torchaudio
numpy
opencv-python
pillow
matplotlib
fastapi
uvicorn
python-multipart
huggingface_hub
requests
